Revert "Deprecate SIP service support."
This reverts commit 912dfd8a867f3d8da5944d90a48daa8d76236b14.

Revert https://android-review.googlesource.com/q/topic:deprecate-sip
partially, aka only this commit.

This does not address any of the issues mentioned in the topic,
but at least it maintains previously working functionality.

In the future, let's revisit this - either try to fix the issues,
or switch to different APIs entirely.

Change-Id: Iad9aa42dbb22cf4960f10f26dcff89e9cbbcf7f0